+++ components/form/FormItemLabel.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import type { ColProps } from '../grid/col';
 import Col from '../grid/col';
 import defaultLocale from '../locale/en_US';
-import { useLocale } from '../locale/LocaleReceiver';
+import useLocale from '../locale/useLocale';
 import type { TooltipProps } from '../tooltip';
 import Tooltip from '../tooltip';
 import type { FormContextProps } from './context';
