+++ components/upload/Upload.tsx
@@ -7,7 +7,7 @@ import { flushSync } from 'react-dom';
 import { ConfigContext } from '../config-provider';
 import DisabledContext from '../config-provider/DisabledContext';
 import defaultLocale from '../locale/en_US';
-import { useLocale } from '../locale/LocaleReceiver';
+import useLocale from '../locale/useLocale';
 import warning from '../_util/warning';
 import type { RcFile, ShowUploadListInterface, UploadChangeParam, UploadFile } from './interface';
 import { UploadProps } from './interface';
